started work on streamlining sitespeed tests
