+++ tests/software.py
@@ -12,6 +12,7 @@ import re
 # https://docs.python.org/3/library/urllib.parse.html
 from urllib.parse import urlparse
 from tests.utils import *
+from tests.sitespeed_base import get_result
 import datetime
 import gettext
 _ = gettext.gettext
@@ -67,165 +68,16 @@ raw_data = {
 }
 
 
-def get_foldername_from_url(url):
-    o = urlparse(url)
-    hostname = o.hostname
-    relative_path = o.path
-
-    test_str = '{0}{1}'.format(hostname, relative_path)
-
-    regex = r"[^a-zA-Z0-9\-\/]"
-    subst = "_"
-
-    # You can manually specify the number of replacements by changing the 4th argument
-    folder_result = re.sub(regex, subst, test_str, 0, re.MULTILINE)
-
-    # NOTE: hopefully temporary fix for "index.html" and Gullspangs-kommun.html
-    folder_result = folder_result.replace('_html', '.html')
-
-    folder_result = folder_result.replace('/', os.sep)
-
-    return folder_result
-
-
-def get_sanitized_file_content(input_filename):
-    # print('input_filename=' + input_filename)
-    lines = list()
-    try:
-        with open(input_filename, 'r', encoding='utf-8') as file:
-            data = file.readlines()
-            for line in data:
-                lines.append(line)
-                # print(line)
-    except:
-        print('error in get_local_file_content. No such file or directory: {0}'.format(
-            input_filename))
-        return '\n'.join(lines)
-
-    test_str = '\n'.join(lines)
-    regex = r"[^a-zåäöA-ZÅÄÖ0-9\{\}\"\:;.,#*\<\>%'&$?!`=@\-\–\+\~\^\\\/| \(\)\[\]_]"
-    subst = ""
-
-    # You can manually specify the number of replacements by changing the 4th argument
-    result = re.sub(regex, subst, test_str, 0, re.MULTILINE)
-
-    json_result = json.loads(result)
-    has_minified = False
-    if 'log' not in json_result:
-        return ''
-    if 'version' in json_result['log']:
-        del json_result['log']['version']
-        has_minified = True
-    if 'browser' in json_result['log']:
-        del json_result['log']['browser']
-        has_minified = True
-    if 'creator' in json_result['log']:
-        del json_result['log']['creator']
-        has_minified = True
-    if 'pages' in json_result['log']:
-        has_minified = False
-        for page in json_result['log']['pages']:
-            keys_to_remove = list()
-            for key in page.keys():
-                if key != '_url':
-                    keys_to_remove.append(key)
-            for key in keys_to_remove:
-                del page[key]
-                has_minified = True
-    if 'entries' in json_result['log']:
-        has_minified = False
-        for entry in json_result['log']['entries']:
-            keys_to_remove = list()
-            for key in entry.keys():
-                if key != 'request' and key != 'response':
-                    keys_to_remove.append(key)
-            for key in keys_to_remove:
-                del entry[key]
-                has_minified = True
-
-            keys_to_remove = list()
-            for key in entry['request'].keys():
-                if key != 'url':
-                    keys_to_remove.append(key)
-            for key in keys_to_remove:
-                del entry['request'][key]
-                has_minified = True
-
-            keys_to_remove = list()
-            for key in entry['response'].keys():
-                if key != 'content' and key != 'headers':
-                    keys_to_remove.append(key)
-            for key in keys_to_remove:
-                del entry['response'][key]
-                has_minified = True
-
-    if has_minified:
-        write_json(input_filename, json_result)
-
-    return result
-
-
-def write_json(filename, data):
-    with open(filename, 'w', encoding='utf-8') as outfile:
-        json.dump(data, outfile)
-
-
 def get_rating_from_sitespeed(url, _local, _):
-    # TODO: CHANGE THIS IF YOU WANT TO DEBUG
-    folder_ending = 'tmp'
-    if use_cache:
-        folder_ending = 'cache'
-
-    result_folder_name = os.path.join(
-        'data', 'results-{0}-{1}'.format(str(uuid.uuid4()), folder_ending))
-    # result_folder_name = os.path.join('data', 'results')
-
-    from tests.performance_sitespeed_io import get_result as sitespeed_run_test
-
+    # We don't need extra iterations for what we are using it for
     sitespeed_iterations = 1
+    sitespeed_arg = '--shm-size=1g -b chrome --plugins.remove screenshot --plugins.remove html --plugins.remove metrics --browsertime.screenshot false --screenshot false --screenshotLCP false --browsertime.screenshotLCP false --chrome.cdp.performance false --browsertime.chrome.timeline false --videoParams.createFilmstrip false --visualMetrics false --visualMetricsPerceptual false --visualMetricsContentful false --browsertime.headless true --browsertime.chrome.includeResponseBodies all --utc true --browsertime.chrome.args ignore-certificate-errors -n {0}'.format(
+        sitespeed_iterations)
+    if 'nt' not in os.name:
+        sitespeed_arg += ' --xvfb'
 
-    sitespeed_arg = '--shm-size=1g -b chrome --plugins.remove screenshot --plugins.remove html --plugins.remove metrics --browsertime.screenshot false --screenshot false --screenshotLCP false --browsertime.screenshotLCP false --chrome.cdp.performance false --browsertime.chrome.timeline false --videoParams.createFilmstrip false --visualMetrics false --visualMetricsPerceptual false --visualMetricsContentful false --browsertime.headless true --browsertime.chrome.includeResponseBodies all --outputFolder {2} --utc true --xvfb --browsertime.chrome.args ignore-certificate-errors -n {0} {1}'.format(
-        sitespeed_iterations, url, result_folder_name)
-    if 'nt' in os.name:
-        sitespeed_arg = '--shm-size=1g -b chrome --plugins.remove screenshot --plugins.remove html --plugins.remove metrics --browsertime.screenshot false --screenshot false --screenshotLCP false --browsertime.screenshotLCP false --chrome.cdp.performance false --browsertime.chrome.timeline false --videoParams.createFilmstrip false --visualMetrics false --visualMetricsPerceptual false --visualMetricsContentful false --browsertime.headless true --browsertime.chrome.includeResponseBodies all --outputFolder {2} --utc true --browsertime.chrome.args ignore-certificate-errors -n {0} {1}'.format(
-            sitespeed_iterations, url, result_folder_name)
-        # sitespeed_arg = '--shm-size=1g -b chrome --plugins.remove screenshot --browsertime.chrome.includeResponseBodies all --outputFolder {2} --utc true --browsertime.chrome.args ignore-certificate-errors -n {0} {1}'.format(
-        #     sitespeed_iterations, url, result_folder_name)
-
-    filename = ''
-    # Should we use cache when available?
-    if use_cache:
-        import engines.sitespeed_result as input
-        sites = input.read_sites('', -1, -1)
-        for site in sites:
-            if url == site[1]:
-                filename = site[0]
-                result_folder_name = filename[:filename.rfind(os.path.sep)]
-
-                if is_file_older_than(filename, cache_time_delta):
-                    filename = ''
-                    continue
-
-                file_created_timestamp = os.path.getctime(filename)
-                file_created_date = time.ctime(file_created_timestamp)
-                print('Cached entry found from {0}, using it instead of calling website again.'.format(
-                    file_created_date))
-                break
-
-    if filename == '':
-        sitespeed_run_test(sitespeed_use_docker, sitespeed_arg)
-
-        website_folder_name = get_foldername_from_url(url)
-
-        filename_old = os.path.join(result_folder_name, 'pages',
-                                    website_folder_name, 'data', 'browsertime.har')
-
-        filename = os.path.join(result_folder_name, 'browsertime.har')
-
-        if os.path.exists(filename_old):
-            os.rename(filename_old, filename)
-            dir_old = os.path.join(result_folder_name, 'pages')
-            shutil.rmtree(dir_old)
+    (result_folder_name, filename) = get_result(
+        url, sitespeed_use_docker, sitespeed_arg)
 
     o = urlparse(url)
     origin_domain = o.hostname
@@ -1031,9 +883,9 @@ def identify_software(filename, origin_domain, rules):
     data = list()
 
     # Fix for content having unallowed chars
-    json_content = get_sanitized_file_content(filename)
-    if True:
-        har_data = json.loads(json_content)
+    with open(filename) as json_input_file:
+        har_data = json.load(json_input_file)
+
         if 'log' in har_data:
             har_data = har_data['log']
         for entry in har_data["entries"]:
